Move EyeLike to Experimental API
