+++ ivy_tests/test_ivy/test_functional/test_core/test_creation.py
@@ -385,44 +385,6 @@ def test_eye(
     )
 
 
-# eye_like
-@handle_test(
-    fn_tree="functional.ivy.eye_like",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_num_dims=2,
-        max_num_dims=2,
-        min_dim_size=1,
-        max_dim_size=5,
-    ),
-    k=helpers.ints(min_value=-10, max_value=10),
-    test_gradients=st.just(False),
-)
-def test_eye_like(
-    *,
-    dtype_and_x,
-    k,
-    test_flags,
-    backend_fw,
-    fn_name,
-    on_device,
-    ground_truth_backend,
-):
-    dtype, x = dtype_and_x
-    helpers.test_function(
-        input_dtypes=dtype,
-        test_flags=test_flags,
-        on_device=on_device,
-        fw=backend_fw,
-        fn_name=fn_name,
-        x=x[0],
-        k=k,
-        dtype=dtype[0],
-        device=on_device,
-        ground_truth_backend=ground_truth_backend,
-    )
-
-
 # from_dlpack
 # TODO: Fix container flag
 @handle_test(
